mediatypes: add ultrahd for 4K releases
This is some prep work for later commits in this series, adding some new
media types which will  allow us to handle 4K releases separately from
Blu-Ray.

Within the DVDProfiler export there are 4 media types exported; DVD,
HDDVD, BluRay, and  UltraHD. As no releases exist which contain a UHD
disc along with a HDDVD, we are left with the following possible
combinations:

  * ULTRAHD - a release containing just a UHD disc
  * ULTRAHD_BLURAY - a release with both a UHD disc and a Blu-ray
  * ULTRAHD_BLURAY_DVD - a release with UHD, Blu-ray and DVD discs

Some initial graphics are included, but these are expected to change.

Signed-off-by: Gavin-John Noonan <mail@gjnoonan.co.uk>