.gitignore: ignore images and debug logs
If you are working on/testing the project and have $getimages set to 2
in your siteconfig it will attempt to grab cover art from the invelos
server, a downside is where it puts them. Let's ignore them.

Also ignore ERROR_LOG and DEBUG_LOG which should never be added into
git.

Signed-off-by: Gavin-John Noonan <mail@gjnoonan.co.uk>