functions.php: remove unused variables
There's quite a lot of unused variables throughout the codebase that can
lead to misunderstanding how it works, removing these variables is fine
and _should_ not break any existing features.

Weirdly the variables insive `find_files_in_set` were typos and the code
will have never worked; Greping the codebase shows that the function is
not used however. This can be removed in another patch.

Signed-off-by: Gavin-John Noonan <mail@gjnoonan.co.uk>