Fix Sonar reported bugs
Running Sonar on the existing code base reports many issues, of which
ten are categorized as bugs by the PHP scanner: mainly the use of some
uninitialized variables. After inspection, it appears they are either
a missing global definition, or a typo due to copy-paste.
This commit fixes all of those ten bugs.